Mudando o tipo de dado das variáveis do DESAFIO 003 (int => float)

+++ Exercise_003.py
@@ -1,6 +1,6 @@
 print("====== DESAFIO 003 ======")
 
-first_number = int(input("Primeiro número "))
-second_number = int(input("Segundo número "))
+first_number = float(input("Primeiro número "))
+second_number = float(input("Segundo número "))
 
 print("A soma é ", first_number + second_number)
